refactor: move incident response constants to a dedicated file for better organization
