+++ plugins/wazuh-analysis/public/groups/security-operations/applications.ts
@@ -5,13 +5,15 @@ import {
   AppUpdater,
 } from '../../../../../src/core/public';
 import {
-  INCIDENT_RESPONSE_ID,
-  INCIDENT_RESPONSE_TITLE,
   IT_HYGIENE_ID,
   IT_HYGIENE_TITLE,
   REGULATORY_COMPLIANCE_ID,
   REGULATORY_COMPLIANCE_TITLE,
 } from './constants';
+import {
+  INCIDENT_RESPONSE_ID,
+  INCIDENT_RESPONSE_TITLE,
+} from './apps/incident-response/constants';
 
 export function getSecurityOperationsApps(updater$?: Subject<AppUpdater>) {
   return [
